requirements.txt
requests
pandas
google-cloud-storage
google-auth
apache-airflow
pyyaml